moved parsed back to TPane namespace
